Move SlidingPanel to SlidingPanel folder
