+++ app/components/SlidingPanel/SlidingPanel.vue
